Update bookSlice for the  CRUD functionality
